refactoring variable and classnames accordingly
